+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Session` interface for handling session on `ServerTransport` interfaces.
 - `ProgressReporter` function type for reporting progress in server primitive operations
+- `ResourceSubscriptionHandler` interface for handling resource subscriptions.
 
 ### Changed
 
@@ -31,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `pkg/servers` package (moved to `servers` for cleaner imports).
 - `Transport` interface (replaced by `ServerTransport` and `ClientTransport`)
 - `ProgressReporter` interface (replaced with function type for simpler progress reporting)
+- `ResourceSubscribedUpdater` interface (replaced with `ResourceSubscriptionHandler`)
 
 ## [0.2.0] - 2024-12-27